Complete Bckp test migration into canonical payment/product/core files
