+++ tests/Feature/Core/TaxRateDeletionValidationFeatureTest.php
@@ -154,4 +154,87 @@ class TaxRateDeletionValidationFeatureTest extends AbstractTestCase
         $response2->assertSessionHas('alert_success');
         $this->assertDatabaseMissing('ip_tax_rates', ['tax_rate_id' => $taxRate->tax_rate_id]);
     }
+
+
+    // Migrated from BckpTaxRateDeletionValidationTest.php
+    public function it_allows_deletion_of_tax_rate_without_references(): void
+    {
+        /* Arrange */
+        $taxRate = $this->seedModel('TaxRate', ['tax_rate_name' => 'Unused Tax Rate']);
+
+        /* Act */
+        $canDelete = $this->service->canDelete($taxRate->tax_rate_id);
+        $blockers  = $this->service->getDeletionBlockers($taxRate->tax_rate_id);
+
+        /* Assert */
+        $this->assertTrue($canDelete);
+        $this->assertEquals(0, $blockers['products']);
+        $this->assertEquals(0, $blockers['invoice_items']);
+        $this->assertEquals(0, $blockers['invoice_tax_rates']);
+        $this->assertEquals(0, $blockers['quote_items']);
+        $this->assertEquals(0, $blockers['quote_tax_rates']);
+    }
+
+    public function it_prevents_deletion_with_invoice_tax_rates(): void
+    {
+        /* Arrange */
+        $taxRate = $this->seedModel('TaxRate');
+        $this->seedModel('InvoiceTaxRate', ['tax_rate_id' => $taxRate->tax_rate_id]);
+
+        /* Act */
+        $canDelete = $this->service->canDelete($taxRate->tax_rate_id);
+
+        /* Assert */
+        $this->assertFalse($canDelete);
+    }
+
+    public function it_prevents_deletion_with_quote_tax_rates(): void
+    {
+        /* Arrange */
+        $taxRate = $this->seedModel('TaxRate');
+        $this->seedModel('QuoteTaxRate', ['tax_rate_id' => $taxRate->tax_rate_id]);
+
+        /* Act */
+        $canDelete = $this->service->canDelete($taxRate->tax_rate_id);
+
+        /* Assert */
+        $this->assertFalse($canDelete);
+    }
+
+    public function it_returns_correct_blocker_counts(): void
+    {
+        /* Arrange */
+        $taxRate = $this->seedModel('TaxRate');
+
+        $this->seedModelMany('Product', 2, ['tax_rate_id' => $taxRate->tax_rate_id]);
+        $this->seedModelMany('InvoiceItem', 3, ['item_tax_rate_id' => $taxRate->tax_rate_id]);
+        $this->seedModelMany('QuoteItem', 1, ['item_tax_rate_id' => $taxRate->tax_rate_id]);
+        /* Act */
+        $blockers = $this->service->getDeletionBlockers($taxRate->tax_rate_id);
+
+        /* Assert */
+        $this->assertEquals(2, $blockers['products']);
+        $this->assertEquals(3, $blockers['invoice_items']);
+        $this->assertEquals(1, $blockers['quote_items']);
+    }
+
+    public function it_allows_deletion_after_all_references_removed(): void
+    {
+        /* Arrange */
+        $taxRate = $this->seedModel('TaxRate');
+        $product = $this->seedModel('Product', ['tax_rate_id' => $taxRate->tax_rate_id]);
+
+        // Initially cannot delete
+        $this->assertFalse($this->service->canDelete($taxRate->tax_rate_id));
+
+        // Remove reference
+        $product->delete();
+
+        /* Act */
+        $canDelete = $this->service->canDelete($taxRate->tax_rate_id);
+
+        /* Assert */
+        $this->assertTrue($canDelete);
+    }
+
 }
